web3auth login working - fix env process
