+++ src/pages/landing-page/Landingpage.scss
@@ -31,7 +31,7 @@
   }
 
   &__sideBar {
-    z-index: 3;
+    z-index: 0;
     color: white;
     padding: 2rem;
     width: 70%;
