--- src/utils/web3auth-service/web3auth.jsx
@@ -1,215 +1,0 @@
-import { ADAPTER_EVENTS, SafeEventEmitterProvider } from "@web3auth/base";
-import { Web3Auth } from "@web3auth/web3auth";
-import { OpenloginAdapter } from "@web3auth/openlogin-adapter";
-import { createContext, FunctionComponent, ReactNode, useCallback, useContext, useEffect, useState } from "react";
-import { CHAIN_CONFIG, CHAIN_CONFIG_TYPE } from "../config/chainConfig";
-import { WEB3AUTH_NETWORK_TYPE } from "../config/web3AuthNetwork";
-import { getWalletProvider, IWalletProvider } from "./walletProvider";
-
-// export interface IWeb3AuthContext {
-//   web3Auth: Web3Auth | null;
-//   provider: IWalletProvider | null;
-//   isLoading: boolean;
-//   user: unknown;
-//   chain: string;
-//   login: () => Promise<void>;
-//   logout: () => Promise<void>;
-//   getUserInfo: () => Promise<any>;
-//   signMessage: () => Promise<any>;
-//   getAccounts: () => Promise<any>;
-//   getBalance: () => Promise<any>;
-//   signTransaction: () => Promise<void>;
-//   signAndSendTransaction: () => Promise<void>;
-// }
-
-export const Web3AuthContext = createContext({
-  web3Auth: null,
-  provider: null,
-  isLoading: false,
-  user: null,
-  chain: "",
-  login: async () => { },
-  logout: async () => { },
-  getUserInfo: async () => { },
-  signMessage: async () => { },
-  getAccounts: async () => { },
-  getBalance: async () => { },
-  signTransaction: async () => { },
-  signAndSendTransaction: async () => { },
-});
-
-export function useWeb3Auth() {
-  return useContext(Web3AuthContext);
-}
-
-// interface IWeb3AuthState {
-//   web3AuthNetwork: WEB3AUTH_NETWORK_TYPE;
-//   chain: CHAIN_CONFIG_TYPE;
-// }
-// interface IWeb3AuthProps {
-//   children?: ReactNode;
-//   web3AuthNetwork: WEB3AUTH_NETWORK_TYPE;
-//   chain: CHAIN_CONFIG_TYPE;
-// }
-
-export const Web3AuthProvider = ({ children, web3AuthNetwork, chain }) => {
-  const [web3Auth, setWeb3Auth] = useState(null);
-  const [provider, setProvider] = useState(null);
-  const [user, setUser] = useState(null);
-  const [isLoading, setIsLoading] = useState(false);
-
-  const setWalletProvider = useCallback(
-    (web3authProvider) => {
-      const walletProvider = getWalletProvider(chain, web3authProvider, uiConsole);
-      setProvider(walletProvider);
-    },
-    [chain]
-  );
-
-  useEffect(() => {
-    const subscribeAuthEvents = (web3auth) => {
-      // Can subscribe to all ADAPTER_EVENTS and LOGIN_MODAL_EVENTS
-      web3auth.on(ADAPTER_EVENTS.CONNECTED, (data) => {
-        console.log("Yeah!, you are successfully logged in", data);
-        setUser(data);
-        setWalletProvider(web3auth.provider);
-      });
-
-      web3auth.on(ADAPTER_EVENTS.CONNECTING, () => {
-        console.log("connecting");
-      });
-
-      web3auth.on(ADAPTER_EVENTS.DISCONNECTED, () => {
-        console.log("disconnected");
-        setUser(null);
-      });
-
-      web3auth.on(ADAPTER_EVENTS.ERRORED, (error) => {
-        console.error("some error or user has cancelled login request", error);
-      });
-    };
-
-    const currentChainConfig = CHAIN_CONFIG[chain];
-
-    async function init() {
-      try {
-        setIsLoading(true);
-        const clientId = "BKPxkCtfC9gZ5dj-eg-W6yb5Xfr3XkxHuGZl2o2Bn8gKQ7UYike9Dh6c-_LaXlUN77x0cBoPwcSx-IVm0llVsLA";
-        const web3AuthInstance = new Web3Auth({
-          chainConfig: currentChainConfig,
-          // get your client id from https://dashboard.web3auth.io
-          clientId,
-        });
-        const adapter = new OpenloginAdapter({ adapterSettings: { network: web3AuthNetwork, clientId } });
-        web3AuthInstance.configureAdapter(adapter);
-        subscribeAuthEvents(web3AuthInstance);
-        setWeb3Auth(web3AuthInstance);
-        await web3AuthInstance.initModal();
-      } catch (error) {
-        console.error(error);
-      } finally {
-        setIsLoading(false);
-      }
-    }
-    init();
-  }, [chain, web3AuthNetwork, setWalletProvider]);
-
-  const login = async () => {
-    if (!web3Auth) {
-      console.log("web3auth not initialized yet");
-      uiConsole("web3auth not initialized yet");
-      return;
-    }
-    const localProvider = await web3Auth.connect();
-    setWalletProvider(localProvider);
-  };
-
-  const logout = async () => {
-    if (!web3Auth) {
-      console.log("web3auth not initialized yet");
-      uiConsole("web3auth not initialized yet");
-      return;
-    }
-    await web3Auth.logout();
-    setProvider(null);
-  };
-
-  const getUserInfo = async () => {
-    if (!web3Auth) {
-      console.log("web3auth not initialized yet");
-      uiConsole("web3auth not initialized yet");
-      return;
-    }
-    const user = await web3Auth.getUserInfo();
-    uiConsole(user);
-  };
-
-  const getAccounts = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.getAccounts();
-  };
-
-  const getBalance = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.getBalance();
-  };
-
-  const signMessage = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.signMessage();
-  };
-
-  const signTransaction = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.signTransaction();
-  };
-
-  const signAndSendTransaction = async () => {
-    if (!provider) {
-      console.log("provider not initialized yet");
-      uiConsole("provider not initialized yet");
-      return;
-    }
-    await provider.signAndSendTransaction();
-  };
-
-  const uiConsole = (...args) => {
-    const el = document.querySelector("#console>p");
-    if (el) {
-      el.innerHTML = JSON.stringify(args || {}, null, 2);
-    }
-  };
-
-  const contextProvider = {
-    web3Auth,
-    chain,
-    provider,
-    user,
-    isLoading,
-    login,
-    logout,
-    getUserInfo,
-    getAccounts,
-    getBalance,
-    signMessage,
-    signTransaction,
-    signAndSendTransaction,
-  };
-  return <Web3AuthContext.Provider value={contextProvider}>{children}</Web3AuthContext.Provider>;
-};
